hybrid autoencoder: split it into multiple files
